DataTypes -> Float in models,minor fixes

+++ back_end_SQL/migrations/20220906131407-create-order.js
@@ -15,7 +15,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       price: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.FLOAT
       },
       portfolioId: {
         type: Sequelize.INTEGER
