+++ back_end_SQL/models/order.js
@@ -10,7 +10,7 @@ module.exports = (sequelize, DataTypes) => {
   Order.init({
     date: DataTypes.INTEGER,
     quantity: DataTypes.INTEGER,
-    price: DataTypes.DECIMAL,
+    price: DataTypes.FLOAT,
     portfolioId:{
       type:DataTypes.INTEGER,
       onDelete:'CASCADE',
